Move StackFrameHelper to its own file
